interactors allocate to modules
