+++ app/interactors/comments/save.rb
@@ -1,0 +1,11 @@
+module Comments
+  class Save
+    include Interactor
+
+    delegate :comment, to: :context
+
+    def call
+      comment.save || context.fail!
+    end
+  end
+end
