+++ app/interactors/comments/submit.rb
@@ -1,0 +1,7 @@
+module Comments
+  class Submit
+    include Interactor::Organizer
+
+    organize Save, Notifications::Notify
+  end
+end
