--- app/interactors/create_comment.rb
@@ -1,5 +1,0 @@
-class CreateComment
-  include Interactor::Organizer
-
-  organize SaveComment, NotifyUser
-end
