--- app/interactors/create_page.rb
@@ -1,5 +1,0 @@
-class CreatePage
-  include Interactor::Organizer
-
-  organize SavePage, AddHashtag
-end
