+++ app/interactors/pages/submit.rb
@@ -1,0 +1,7 @@
+module Pages
+  class Submit
+    include Interactor::Organizer
+
+    organize Save, Hashtags::Add
+  end
+end
