--- app/interactors/save_comment.rb
@@ -1,9 +1,0 @@
-class SaveComment
-  include Interactor
-
-  delegate :comment, to: :context
-
-  def call
-    comment.save || context.fail!
-  end
-end
