--- app/interactors/save_page.rb
@@ -1,9 +1,0 @@
-class SavePage
-  include Interactor
-
-  delegate :page, to: :context
-
-  def call
-    page.save || context.fail!
-  end
-end
